Squash karel/datasets into karel/dataset, move scripts & notebooks into karel folder
