--- program_synthesis/karel/data/karel/word.vocab
@@ -1,43 +1,0 @@
-<S>	0
-</S>	1
-<UNK>	2
-|||	3
-move	4
-c(	5
-c)	6
-putMarker	7
-DEF	8
-m)	9
-m(	10
-run	11
-pickMarker	12
-turnLeft	13
-i)	14
-i(	15
-turnRight	16
-REPEAT	17
-r(	18
-r)	19
-w)	20
-w(	21
-WHILE	22
-e)	23
-IFELSE	24
-e(	25
-ELSE	26
-IF	27
-leftIsClear	28
-rightIsClear	29
-frontIsClear	30
-markersPresent	31
-noMarkersPresent	32
-not	33
-R=5	34
-R=2	35
-R=4	36
-R=3	37
-R=6	38
-R=7	39
-R=8	40
-R=9	41
-R=10	42
