+++ program_synthesis/karel/dataset/coverage.py
@@ -1,5 +1,5 @@
-from program_synthesis.karel.datasets import executor
-from program_synthesis.karel.datasets.karel import parser_for_synthesis
+from program_synthesis.karel.dataset import executor
+from program_synthesis.karel.dataset import parser_for_synthesis
 
 branch_types = {'if', 'ifElse', 'while'}
 stmt_types = {'move', 'turnLeft', 'turnRight', 'putMarker', 'pickMarker'}
