--- program_synthesis/karel/datasets/__init__.py
@@ -1,5 +1,0 @@
-from program_synthesis.datasets.dataset import get_train_dataset
-from program_synthesis.datasets.dataset import get_eval_dataset
-from program_synthesis.datasets.dataset import get_eval_final_dataset
-from program_synthesis.datasets.dataset import CodeExample
-from program_synthesis.datasets.dataset import set_vocab
