--- program_synthesis/karel/datasets/karel/__init__.py
@@ -1,17 +1,0 @@
-from program_synthesis.datasets.karel.parser_for_synthesis import KarelForSynthesisParser
-from program_synthesis.datasets.karel.utils import  KarelSyntaxError
-from program_synthesis.datasets.karel.utils import  TimeoutError
-from program_synthesis.datasets.karel.utils import  beautify
-from program_synthesis.datasets.karel.utils import  makedirs
-from program_synthesis.datasets.karel.utils import  pprint
-from program_synthesis.datasets.karel.utils import str2bool
-
-from datetime import datetime
-
-__copyright__ = 'Copyright 2015 - {} by Taehoon Kim'.format(datetime.now().year)
-__version__ = '1.3.0'
-__license__ = 'BSD'
-__author__ = 'Taehoon Kim'
-__email__ = 'carpedm20@gmail.com'
-__source__ = 'https://github.com/carpedm20/karel'
-__description__ = 'Karel dataset for program synthesis and program induction'
