+++ program_synthesis/karel/models/modules/karel_edit.py
@@ -6,8 +6,8 @@ import torch
 from torch import nn
 from torch.autograd import Variable
 
-from program_synthesis.karel.datasets import data
-from program_synthesis.karel.datasets.karel import mutation
+from program_synthesis.karel.dataset import data
+from program_synthesis.karel.dataset import mutation
 from program_synthesis.karel.models import beam_search
 from program_synthesis.karel.models.modules import attention
 from program_synthesis.karel.models.modules import karel
