+++ program_synthesis/karel/scripts/split_indexed.py
@@ -4,7 +4,8 @@ import struct
 
 import tqdm
 
-from program_synthesis.datasets import indexed_file
+from program_synthesis.tools import indexed_file
+
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
