--- scripts/setup.py
@@ -1,12 +1,0 @@
-from setuptools import setup
-
-
-VERSION = '0.1.0'
-
-setup(
-    name='program_synthesis',
-    version=VERSION,
-    description='NEAR Program Synthesis',
-    packages=['program_synthesis'],
-    author='NEAR Inc & Contributors'
-)
